+++ src/acquirium/__init__.py
@@ -1,4 +1,5 @@
 from acquirium.Client.acquirium import Acquirium
+from acquirium.runtime import init, shutdown
 from acquirium.Materialization import App, OutputSpec, align, output
 from acquirium.Client.explore.core import Query
 from acquirium.Client.query import Q